+++ src/test/kotlin/nl/mvdr/adventofcode/adventofcode2021/day14/ExtendedPolymerizationPart2Test.kt
@@ -3,7 +3,7 @@ package nl.mvdr.adventofcode.adventofcode2021.day14
 import nl.mvdr.adventofcode.FunctionSolverTest
 import org.junit.jupiter.params.provider.Arguments
 
-class ExtendedPolymerizationPart2Test: FunctionSolverTest<Int>(::solvePart2) {
+class ExtendedPolymerizationPart2Test: FunctionSolverTest<Long>(::solvePart2) {
     companion object {
         @JvmStatic
         fun testSolution(): List<Arguments> = listOf(
